Move ts components to root
