--- react/Border.js
@@ -1,3 +1,0 @@
-import Border from './components/Border'
-
-export default Border
