+++ react/Border.tsx
@@ -1,7 +1,8 @@
-import React, { FunctionComponent } from 'react'
+import type { FunctionComponent } from 'react'
+import React from 'react'
 import { useCssHandles } from 'vtex.css-handles'
 
-import StickyRow from './StickyRow'
+import StickyRow from './components/StickyRow'
 
 interface Props {
   sticky: boolean
@@ -11,6 +12,7 @@ const CSS_HANDLES = ['headerBorder'] as const
 
 const Border: FunctionComponent<Props> = ({ sticky }) => {
   const handles = useCssHandles(CSS_HANDLES)
+
   return (
     <StickyRow sticky={sticky}>
       <div className={`${handles.headerBorder} bb b--muted-3`} />
