--- react/Layout.js
@@ -1,3 +1,0 @@
-import Layout from './components/Layout'
-
-export default Layout
