+++ react/Layout.tsx
@@ -1,6 +1,7 @@
-import React, { FunctionComponent } from 'react'
+import type { FunctionComponent } from 'react'
+import React from 'react'
 
-import StickyRows from './StickyRows'
+import StickyRows from './components/StickyRows'
 
 const Layout: FunctionComponent = ({ children }) => {
   return <StickyRows>{children}</StickyRows>
