--- react/Row.js
@@ -1,3 +1,0 @@
-import Row from './components/Row'
-
-export default Row
