--- react/Spacer.js
@@ -1,3 +1,0 @@
-import Spacer from './components/Spacer'
-
-export default Spacer
